requirements.txt
torch==2.8.0
torchvision==0.23.0
flash_attn==2.7.4.post1
diffusers==0.34.0
transformers==4.53.2
accelerate
tokenizers==0.21.4
safetensors
sentencepiece
protobuf
Pillow
runpod
huggingface_hub
imageio
imageio-ffmpeg
numpy==1.26.4
einops
easydict
ftfy
tqdm
opencv-python-headless
kornia
omegaconf
torchao==0.10.0
pyloudnorm
librosa
moviepy==2.2.1
